Attempt to fix the public event page
