+++ src/app/(pages)/events/public/page.tsx
@@ -1,0 +1,7 @@
+import PublicEvent from "@/publicEvent/templates/PublicEvent";
+import React from "react";
+const page = () => {
+  return <PublicEvent />;
+};
+
+export default page;
