+++ src/publicEvent/templates/PublicEvent.tsx
@@ -1,0 +1,17 @@
+import type React from "react";
+import { Suspense } from "react";
+import Header from "../../home/organisms/Header";
+import EventSearchNavigation from "../organisms/EventSearchNavigation";
+
+const PublicEvent: React.FC = () => {
+  return (
+    <main className="bg-black px-16 text-white min-h-screen">
+      <Header />
+      <Suspense fallback={<div>Loading events...</div>}>
+        <EventSearchNavigation />
+      </Suspense>
+    </main>
+  );
+};
+
+export default PublicEvent;
